调整逻辑

+++ src/main/java/com/science/gtnl/common/machine/multiblock/wireless/SmeltingMixingFurnace.java
@@ -311,7 +311,7 @@ public class SmeltingMixingFurnace extends WirelessEnergyMultiMachineBase<Smelti
     @Override
     public final void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ,
         ItemStack aTool) {
-        this.machineMode = (byte) ((this.machineMode + 1) % 2);
+        this.machineMode = (this.machineMode + 1) % 2;
         GTUtility.sendChatToPlayer(
             aPlayer,
             StatCollector.translateToLocal("SmeltingMixingFurnace_Mode_" + this.machineMode));
